Combination of "yield" and "return" yields unexpected results
Suppose we have a PredicateExpression "p(a,b)" and call .symbols(). This should generate the list ["p", "a", "b"], but only generates ["p"].
This is because "yield" is not supposed to be used together with "return" (https://stackoverflow.com/q/16780002/). Instead, only either "yield" or "return" (as proposed here) should be used.